♻️ #299 Refactor: 폴더 및 파일 구조를 개선하라
